Image of the beehive scale setup
The setup comprises: two sensor clients (TTGO LoRa32), and a LoRa-to-serial interface (TTGO LoRa32) which is connected to a Raspberry Pi 3B+.
A python server collects the data from each client and saves it to a CSV file.  The data viewing software Kst (running on the Pi) is optimal for viewing long term and realtime measurements.